[AOTI][refactor] Move generate_user_defined_triton_kernel
Summary: Move generate_user_defined_triton_kernel from cpp_wrapper_cpu to cpp_wrapper_cuda

[ghstack-poisoned]
