+++ torch/_inductor/codegen/cpp_wrapper_cuda.py
@@ -77,6 +77,36 @@ class CppWrapperCuda(CppWrapperCpu):
             self.prefix.writeline("\n")
         return super().generate(is_inference)
 
+    def generate_user_defined_triton_kernel(
+        self, kernel_name, grid, configs, args, triton_meta, raw_args
+    ):
+        assert len(grid) != 0
+        if len(grid) == 1:
+            grid_decision = grid[0]
+        else:
+            meta = CudaKernelParamCache.get(kernel_name)
+            assert meta is not None
+            grid_decision = None
+            for i, c in enumerate(configs):
+                if all(arg == meta["meta"][key] for key, arg in c.kwargs.items()):
+                    grid_decision = grid[i]
+                    break
+            assert grid_decision is not None
+
+        arg_types = [
+            arg.get_dtype() if hasattr(arg, "get_dtype") else type(arg)
+            for arg in raw_args
+        ]
+        self.generate_kernel_call(
+            kernel_name,
+            args,
+            arg_types=arg_types,
+            grid=grid_decision,
+            cuda=True,
+            triton=True,
+            triton_meta=triton_meta,
+        )
+
     @functools.lru_cache(None)  # noqa: B019
     def generate_load_kernel_once(
         self,
